Moving Folders
Changes to components/c4gt22
- Added contributors2022.js, mentors22.js, projects22.js
- Modified imports to mentorCard and projectCard folders

Changes to components/c4gt23
- added folder mentoringProgram
      - moved in productCard, productList, searchBar folder
      - added contributors2023.js, projects23.js

Added  components/communityProgram folder
- moved in projectTable folder

Added components/home folder
- moved in homeBody and homeHeader folder

import changes to components/index.tsx
