+++ src/components/communityProgram/projectTable/appliedFilters.tsx
